Sharing SwiftUI state with @StateObject
In order to share swiftUI data between views with classes first, we need to add @Published property observers to the properties that we want to be observed when they changed. Second, we need to add @StateObject property wrapper to tell SwiftUI that we're creating a new class instance that should be watched for any change. Finally @StateObject property wrapper can only be used if our class conforms to  ObservableObject protocol.