Showing and hiding views(using sheet view)
One of the ways to show views in SwiftUI is using sheet view. With sheet we can present a view on top of our existing one. First, we create a second view we want to use as a sheet. Second we add state to track if sheet being showed or not. We add toggle to button to change the state of sheet. Third we attach our sheet to a button, created in our ContentView. Fourth we add the view that we want to present to sheet.

To dismiss the second view we can use @Environment property wrapper. We can add dismiss property with @Environment property wrapper which will allow us to store values provided externally. Then we can create a button that calls the DismissAction.