Deleting items using onDelete()
onDelete() modifier can be used to delete an object from a collection. We can use onDelete() modifier with List and ForEach. If we want to use onDelete() with List we have to attach it to ForEach.

In order to make onDelete() work, we need to implement a method with parameter of type IndexSet. It is for tracking the positions of the items in the ForEach. Then we can modify ForEach to perform method that we implemented which will delete data from ForEach.

We can add Edit/Done Button to Navigation Bar to delete rows. To do this we need to wrap our view in NavigationView and add toolbar modifier with EditButton.