Storing user settings with UserDefaults
We can store user data in several ways. One of them is UserDefaults. The data stored in UserDefaults will be loaded when app launches.

We need to use UserDefaults.standard, it is the built-in instance of UserDefaults with set() method that accepts data like booleans, integers and strings. Then we attach a string name as key. We also need to change the property that we want to read with UserDefaults.

We can also use @AppStorage property wrapper to store user defaults. It will work like @State property wrapper. When value changed it will reinvoked body property to UI reflects the new data.