Archiving Swift objects with Codable
We can use Codable protocol for archiving or unarchiving data. It will be useful to store complex data like custom swift types.

We can use JSONEncoder with types that conforms Codable protocol which will take and send back the object in JSON (JavaScript Object Notation). To convert our data to JSON we need to use encode() method on a JSONEncoder. After that we have JSON and we want to convert it to swift Codable types, JSONDecoder should be used.