Sharing an observed object with a new view
Classes that conform ObservableObject protocol can be used more than one SwiftUI view, and all of those views will be updated when published property of the class changes.

Add a SwiftUI View called AddView. Add two TextField for name and amount and a Picker for type inside a From and NavigationView.

Add showingAddExpense property to ContentView to track AddView is being shown or not. Attach sheet modifier to List to present AddView as a sheet.

Not to create second instance of Expenses class in AddView, we should add expenses property with @ObservedObject property wrapper. This will allow us to pass our existing object from one view to another. Views will both share the same object and will monitor it for changes.

Modify button in ContentView to show AddView view as sheet.