Final polish
In order to dismiss AddView after user adds an expense, we add dismiss property with environment property wrapper and call dismiss when we use Save button. This will cause showingAddExpense boolean in ContentView to be false and hide AddView.

To add details for every expense item we can add VStack inside a HStack with name, type and amount Text.