Resizing images to fit the screen using GeometryReader
Add an image with a frame. To resize image contents we can use resizable() modifier. If we use only resizable() our image can be distorted. To resize image proportionally we can use scaleToFit() and scaleToFill() modifiers.

scaleToFit() modifier will make image fit inside the container. It can leave some parts of the view empty.

scaleToFill() modifier will make image fill the all view without any empty space. It can leave some parts of image outside the container.

If we want images automatically scale up to fill more of the screen in one or more dimensions and not to add width or height hard-coded, we can use GeometryReader.

When we use GeometryReader view, we will be able to use GeometryProxy object. We can make  image that's %80 the width of the screen. Also if we want to center a view inside GeometryReader, we can add a second frame that makes it
fill the full space of the container.