How ScrollView lets us work with scrolling data
Add a list of 100 text views with scrollView. ScrollViews can be horizontally, vertically and in both directions. We can also decide the system should show scroll indicators next to them.

Add CustomText wrapper to demonstrate the views that added to scroll view are created immediately. For every item added we will se a message on Xcode's log. If we don't want to create views until they are actually shown and minimize the amount of system resources being use, we can use LazyVStack and LazyHStack.

There is a difference between regular stacks and lazy stacks which is lazy stacks always take up as much as room as is available in our layouts, whereas regular stacks take up only as much space as needed. Because it stops lazy stacks having to adjust their size if a new view is loaded that wants more space.