Pushing new views onto the stack using NavigationLink
SwiftUI's NavigationView shows a navigation bar at the top of the views and it let us push views into a view stack.

Add a list with 100 text items within NavigationView. All rows of the list can be tapped to show a detail view with NavigationLink. We can use NavigationLink with any kind of view.