How to lay out views in a scrolling grid
If want to have grid of information in our views we can use LazyHGrid for horizontal and LazyVGrid for vertical data. Lazy is not to use too much system resources. Views will be created when they needed.

Define grid and place that grid inside a ScrollView. If we want to use the grid with different screen sizes we can do this by using adaptive sizes when we create a GridItem. we can specify maximum and minimum ranges for columns or rows.